% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.build_query_string}
\alias{.build_query_string}
\title{.build_query_string}
\usage{
.build_query_string(args)
}
\arguments{
\item{args}{list of parameters}
}
\value{
character
}
\description{
Builds a query string ready for been added to a url, from a list of name/value parameters
}
\examples{
\dontrun{
.build_query_string(list(name="Bob", city="Berlin"))
}

}
