% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api_tools.R
\name{.build_uri}
\alias{.build_uri}
\title{.build_uri}
\usage{
.build_uri(endpoint, ..., query = list(), api_base = NULL)
}
\arguments{
\item{endpoint}{Name of an endopoint. The endopoint must have been configured
previously with .setup_api_endpoint}

\item{...}{Custom parameters for the query string}

\item{query}{URI parameters for the query string passed as a named list}

\item{api_base}{API base for this URI, no matter if global value was defined}
}
\value{
character
}
\description{
Generates a URI based on the endpoint configuration and query parameters
}
