% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.collapse_array_columns_map}
\alias{.collapse_array_columns_map}
\title{.collapse_array_columns_map}
\usage{
.collapse_array_columns_map(element)
}
\arguments{
\item{reg_list}{data rows as a list of lists}
}
\value{
dataframe
}
\description{
Maps multivalue elements to comma separated strings
}
