% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.extract_response_body}
\alias{.extract_response_body}
\title{.extract_response_body}
\usage{
.extract_response_body(response)
}
\arguments{
\item{response}{Raw http response}
}
\value{
character
}
\description{
Extracts de body from a HTTP response or throws error if not 200 status
}
