% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.get_data_from_uri}
\alias{.get_data_from_uri}
\title{.get_data_from_uri}
\usage{
.get_data_from_uri(uri)
}
\arguments{
\item{uri}{Public internet address of the data}
}
\value{
dataframe with rows of data
}
\description{
Grabs data as dataframe from a URI. Expects the response data to be in JSON format
}
\examples{
\dontrun{
.get_data_from_uri("http://api.example.com/ecologydata")
}
}
