% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_querys.R
\name{.implode_to_string}
\alias{.implode_to_string}
\title{.implode_to_string}
\usage{
.implode_to_string(params)
}
\arguments{
\item{params}{list of strings}
}
\value{
character
}
\description{
Converts a list of strings in a single comma separated string
}
\examples{
\dontrun{
.implode_to_string(list("categoryA","categoryB","categoryC"))
}
}
