% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.make_data_frame}
\alias{.make_data_frame}
\title{.make_data_frame}
\usage{
.make_data_frame(reg_list)
}
\arguments{
\item{reg_list}{data rows as a list of lists}
}
\value{
dataframe
}
\description{
Makes a dataframe from a list of lists
}
