% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_cache.R
\name{.package_cache_get}
\alias{.package_cache_get}
\title{.package_cache_get}
\usage{
.package_cache_get(varname)
}
\arguments{
\item{varname}{Name of the variable}
}
\value{
value of the variable
}
\description{
Retrieves the value of a cache variable already initialized
}
