% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_cache.R
\name{.package_cache_has}
\alias{.package_cache_has}
\title{.package_cache_has}
\usage{
.package_cache_has(varname)
}
\arguments{
\item{varname}{Name of the variable}
}
\value{
boolean
}
\description{
Check weather a variable exists in the package cache environment
}
