% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_cache.R
\name{.package_cache_return_or_setup}
\alias{.package_cache_return_or_setup}
\title{.package_cache_return_or_setup}
\usage{
.package_cache_return_or_setup(varname, setupfun)
}
\arguments{
\item{varname}{Name of the variable to retrieve}

\item{setupfun}{Function to compute the value of the variable in case it is not yet initialized}
}
\value{
value of the cached variable
}
\description{
Returns a cached variable value. If it's intializated already, it will be initialized by calling the function provided
}
\examples{
\dontrun{
.package_cache_return_or_setup("specieslist", function(){
    list_from_remote_server <- list("dog", "cat", "chupacabra")
    return (list_from_remote_server)
})
}
}
