% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_cache.R
\name{.package_cache_set}
\alias{.package_cache_set}
\title{.package_cache_set}
\usage{
.package_cache_set(varname, value)
}
\arguments{
\item{varname}{Name of the variable}
}
\value{
value of the variable
}
\description{
Sets a variable value in cache
}
\examples{
\dontrun{
.package_cache_set("specieslist", list("dog", "cat", "chupacabra"))
}
}
