% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{.parse_raw_data}
\alias{.parse_raw_data}
\title{.parse_raw_data}
\usage{
.parse_raw_data(raw_data)
}
\arguments{
\item{raw_data}{json encoded data}
}
\value{
dataframe
}
\description{
Parse raw json string as a dataframe
}
