% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_querys.R
\name{.pbdb_query}
\alias{.pbdb_query}
\title{.pbdb_query}
\usage{
.pbdb_query(endpoint, query)
}
\arguments{
\item{endpoint}{Name of the endpoint, inside the API, to which the query must be sent. 
This endpoint must have been previously configured}

\item{query}{List of filter parameters for the api. The values provided to the parameters
may be a single string or a list of strings.}
}
\value{
dataframe
}
\description{
Central function for sending all queries to the remote API
}
\examples{
\dontrun{
.pbdb_query("occs/list", list(base_name="Canidae", show=c("coords", "phylo", "ident")))
}

}
