% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_rest_api_setup.R
\name{.pbdb_set_up_endpoints}
\alias{.pbdb_set_up_endpoints}
\title{.pbdb_set_up_endpoints}
\usage{
.pbdb_set_up_endpoints()
}
\description{
This function registers all the endpoints available from the paleobiodb.org REST API
}
