% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_rest_api_setup.R
\name{.pbdb_uri_builder}
\alias{.pbdb_uri_builder}
\title{.pbdb_uri_builder}
\usage{
.pbdb_uri_builder(api_base_url, config, querystring = "")
}
\arguments{
\item{api_base_url}{Base url for the API}

\item{config}{Configuration of the endpoint}

\item{querystring}{compiled querystring}
}
\description{
Function that generates the URIs for the paleobiodb.org API
}
