% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_cache.R
\docType{data}
\name{.private_pk_cache}
\alias{.private_pk_cache}
\title{.private_pk_cache}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.private_pk_cache
}
\description{
Environment for the runtime variables of the package
}
\keyword{datasets}
