% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api_tools.R
\name{.setup_api_endpoint}
\alias{.setup_api_endpoint}
\title{.setup_api_endpoint}
\usage{
.setup_api_endpoint(
  name,
  endpoint_base,
  ...,
  uri_builder = .default_uri_builder,
  query_params = list()
)
}
\arguments{
\item{name}{Internal use name of the endpoint}

\item{endpoint_base}{Base of the endpoint. Should not include de api_base. 
eg: products in a http://sample.org/api/products where http://sample.org/api 
ould be the api_base}

\item{...}{Custom parameters passed to the uri_builder callback when generating the URIs}

\item{uri_builder}{Function that generates the URIs to the API}
}
\description{
Sets up configuration for an API endpoint
}
