% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api_tools.R
\name{.stop_on_missing_param}
\alias{.stop_on_missing_param}
\title{.stop_on_missing_param}
\usage{
.stop_on_missing_param(compulsory_params, query)
}
\arguments{
\item{compulsory_params}{list of compulsory}

\item{query}{list of query name/value pairs}
}
\description{
Throws error if a compulsory parameter is not found in query
}
