\docType{package}
\name{pbdb}
\alias{pbdb}
\alias{pbdb-package}
\title{pbdb: An R-package to download the information stored in Paleobiology Database,}
\description{
R wrapper to the Paleobiology Database.
}
\details{
\tabular{ll}{ Package: \tab pbdb \cr Type: \tab Package\cr
Version: \tab 0.1\cr Date: \tab 2014-02-15\cr License: \tab
GPL-2 \cr }

We programmed functions to set flexible queries about the
species occurrences...blabla
}
\examples{
\dontrun{
blabla
}
}
\author{
Javier Gonzalez \email{javigzz@yahoo.es}

Sara Varela \email{svarela@paleobiogeography.org}
}
\references{
in prep.
}
\seealso{
{ http://paleobiodb.org }
}
\keyword{package}

