% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pbdb_export_PyRate}
\alias{pbdb_export_PyRate}
\title{pbdb_export_PyRate}
\usage{
pbdb_export_PyRate (base_name)
}
\arguments{
\item{base_name}{Return only records associated with the specified taxonomic name(s), or any of their children. e.g. Canidae; the function will download all the occurrences from the Paleobiology Database
of all the species belonging to the Canidae family.}
}
\value{
a dataframe with the species, minimum and maximum age of the occurrences, and of the species is extinct or extant.
}
\description{
Returns a dataframe with the adequate structure to run PyRate
http://www2.unil.ch/phylo/bioinformatics/pyrate.html
}
\examples{
\dontrun{

pbdb_export_PyRate ("Hyaenidae")

}
}

