% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_queries.R
\name{pbdb_opinions_taxa}
\alias{pbdb_opinions_taxa}
\title{Get taxonomic opinions about taxa}
\usage{
pbdb_opinions_taxa(...)
}
\arguments{
\item{...}{Arguments passed to the API. See documentation for
accepted parameters at
\url{https://paleobiodb.org/data1.2/taxa/opinions}.}
}
\value{
A data frame with information about the taxonomic opinions
that match the query.
}
\description{
Returns information about the taxonomic opinions used to build the
taxonomic hierarchy. From all of the opinions entered into the
database about a particular taxon, the most recent opinion that is
stated with the most evidence is used to classify that taxon. The
others are considered to be superseded and are ignored.
}
\examples{
\dontrun{
  pbdb_opinions_taxa(base_name = "Canis")
}
}
