% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdb_querys.R
\name{pbdb_scale}
\alias{pbdb_scale}
\title{pbdb_scale}
\usage{
pbdb_scale (id, ...)
}
\arguments{
\item{id}{identifier of the temporal interval.
This parameter is required.}

\item{...}{documentation for all the parameters
is available in http://paleobiodb.org/data1.1/scales/single.
For example:
\code{vocab} set vocab="pbdb" to show the complete name
of the variables
(by default variables have short 3-letter names)}
}
\value{
a dataframe with information from a single scale
}
\description{
Returns information about a single time scale, selected by
identifier.
}
\examples{
\dontrun{
pbdb_scale (id=1, vocab="pbdb")
}
}

