% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_queries.R
\name{pbdb_scales}
\alias{pbdb_scales}
\title{Get information about multiple time scales}
\usage{
pbdb_scales(...)
}
\arguments{
\item{...}{Arguments passed to the API. See
documentation for accepted parameters at
\url{https://paleobiodb.org/data1.2/timescales/list}. E.g.:
\itemize{
\item \code{all_records}: Set to \code{TRUE} to list all intervals.
\item \code{id}: Return only time scales with the specified
identifier(s). It is possible to provide multiple identifiers as
a comma-separated list.
\item \code{name}: Return only time scales with the specified name(s). It is
possible to provide multiple names as a comma-separated list.
\item \code{vocab}: Set to \code{"pbdb"} to show the complete name of the
variables (by default variables have short 3-letter names).
}}
}
\value{
A data frame with information from the selected scales.
}
\description{
Returns information about multiple time scales.
}
\examples{
\dontrun{
  ## Get a data frame with all the scales available in PBDB
  ## by setting no ids
  pbdb_scales(all_records = TRUE)
}
}
