% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdb_querys.R
\name{pbdb_taxa_auto}
\alias{pbdb_taxa_auto}
\title{pbdb_taxa_auto}
\usage{
pbdb_taxa_auto (...)
}
\arguments{
\item{...}{see all paramenters at http://paleobiodb.org/data1.1/taxa/auto_doc.html}

\item{name}{A partial name or prefix.
It must have at least 3 significant characters, and may include both a genus
(possibly abbreviated) and a species.}

\item{limit}{set the limit to the number of matches}
}
\value{
a dataframe with information about the matches
(taxon rank and number of occurrences in the database)
}
\description{
Returns a list of names matching the given prefix or partial name.
}
\details{
See all paramenters at:
http://paleobiodb.org/data1.1/taxa/auto_doc.html
}
\examples{
\dontrun{
pbdb_taxa_auto (name="Cani", limit=10)
}
}

