% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_temporal_functions.R
\name{pbdb_temp_range}
\alias{pbdb_temp_range}
\title{Temporal range of taxa}
\usage{
pbdb_temp_range(data, rank, col = "#0000FF", names = TRUE, do.plot = TRUE)
}
\arguments{
\item{data}{Data frame from a query to PaleobioDB as returned by
\code{\link[=pbdb_occurrences]{pbdb_occurrences()}}.  Important: it is required to
show the name of the families, orders, etc. in the data frame, to
do that set: \code{show = "classext"} or \code{show = "class"}
(see Examples).}

\item{rank}{The taxon rank to be analyzed.}

\item{col}{Colour of the bars in the plot.}

\item{names}{Logical indicating whether to include the name of the
taxa in the plot (\code{TRUE} by default).}

\item{do.plot}{Logical value indicating whether to produce a plot
(\code{TRUE} by default).}
}
\value{
A data frame with the time span of the taxa selected
(species, genera, etc.).
}
\description{
Returns a data frame with the temporal range of the taxa within a
selected rank (species, genera, families, etc.), and optionally
generates a plot from it.
}
\examples{
\dontrun{
  canis_quaternary <- pbdb_occurrences(
    limit = "all", base_name = "Canis", interval = "Quaternary",
    show = c("coords", "classext", "ident"), vocab = "pbdb"
  )
  pbdb_temp_range(canis_quaternary, rank = "species", names = FALSE)
}
}
