% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_queries.R
\name{pbdb_collections}
\alias{pbdb_collections}
\title{pbdb_collections}
\usage{
pbdb_collections(...)
}
\arguments{
\item{...}{documentation for all the parameters is available in
\url{https://paleobiodb.org/data1.2/colls/list}. Go to
\code{\link{pbdb_occurrences}} to see an explanation about the
main filtering parameters.}
}
\value{
A data frame with the collections that match the query
}
\description{
Returns information about multiple collections, selected according
to the parameters you provide.
}
\examples{
\dontrun{
  pbdb_collections(base_name = "Cetacea", interval = "Miocene")
}
}
