% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pbdb_querys.R
\name{pbdb_occurrence}
\alias{pbdb_occurrence}
\title{pbdb_occurrence}
\usage{
pbdb_occurrence (id, ...)
}
\arguments{
\item{id}{identifier of the occurrence. This parameter is
required.}

\item{...}{see http://paleobiodb.org/data1.1/occs/single for additional filters}

\item{vocab}{set vocab="pbdb" to show the complete name
of the variables (by default variables have short 3-letter names)}

\item{show}{to show extra variables (e.g. coords)}
}
\value{
a dataframe with a single occurrence
}
\description{
Returns information about a single occurrence record from the Paleobiology
Database.
}
\details{
Documentation for all the parameters is available at
http://paleobiodb.org/data1.1/occs/single. In the parameter list above, we
describe the most common filters that paleontologists and ecologists might
use.
}
\examples{
\dontrun{
pbdb_occurrence (id=1001)
pbdb_occurrence (id=1001, vocab="pbdb", show="coords")
}
}

