% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_queries.R
\name{pbdb_reference}
\alias{pbdb_reference}
\title{pbdb_reference}
\usage{
pbdb_reference(id, ...)
}
\arguments{
\item{id}{identifier of the reference. This parameter is required.}

\item{...}{arguments passed to the API. See documentation for
accepted parameters in
\url{https://paleobiodb.org/data1.2/refs/single}. E.g.:
\itemize{
  \item \emph{vocab}: set vocab = "pbdb" to show the complete
    name of the variables (by default variables have short
    3-letter names).
  \item ...
}}
}
\value{
a data frame with a single reference
}
\description{
Returns information about a single reference, selected by
identifier.
}
\examples{
\dontrun{
  pbdb_reference(id = 1003, vocab = "pbdb", show = "both")
}
}
