% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_querys.R
\name{pbdb_references}
\alias{pbdb_references}
\title{pbdb_references}
\usage{
pbdb_references (...)
}
\arguments{
\item{...}{arguments passed to the API. See
documentation for accepted parameters in
\url{http://paleobiodb.org/data1.1/refs/list}. Eg:
\itemize{
  \item \emph{author} select only references for which any of the authors 
    matches the specified name
  \item \emph{year} select only references published in the specified year
  \item \emph{pubtitle} select only references that involve the specified 
    publication
  \item \emph{order} specifies the order in which the results are returned. You can
    specify multiple values separated by commas, and each value may be appended
    with .asc or .desc. Accepted values are: author, year, pubtitle, created,
    modified, rank.
  \item ...
}}
}
\value{
a dataframe with the information about the references that match the query
}
\description{
Returns information about multiple references, selected according to the parameters you provide.
}
\examples{
\dontrun{
  pbdb_references (author= "Polly")
}
}
