% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_queries.R
\name{pbdb_scale}
\alias{pbdb_scale}
\title{pbdb_scale}
\usage{
pbdb_scale(id, ...)
}
\arguments{
\item{id}{identifier of the temporal interval. This parameter is required.}

\item{...}{additional arguments passed to the API. See
documentation for accepted parameters in
\url{https://paleobiodb.org/data1.2/scales/single}. E.g.:
\itemize{
  \item \emph{vocab}: set vocab = "pbdb" to show the complete
    name of the variables (by default variables have short
    3-letter names).
  \item ...
}}
}
\value{
a dataframe with information from a single scale
}
\description{
Returns information about a single time scale, selected by
identifier.
}
\examples{
\dontrun{
  pbdb_scale(id = 1, vocab = "pbdb")
}
}
