% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_queries.R
\name{pbdb_scales}
\alias{pbdb_scales}
\title{pbdb_scales}
\usage{
pbdb_scales(...)
}
\arguments{
\item{...}{arguments passed to the API. See
documentation for accepted parameters in
\url{https://paleobiodb.org/data1.2/scales/list}. E.g.:
\itemize{
  \item \emph{vocab}: set vocab = "pbdb" to show the complete
    name of the variables (by default variables have short
    3-letter names).
  \item ...
}}
}
\value{
a dataframe with information from the selected scales
}
\description{
Returns information about multiple time scales.
}
\examples{
\dontrun{
  ## Get a dataframe with all the scales available in PBDB
  ## by setting no ids
  pbdb_scales()
}
}
