% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_taxonomic_functions.R
\name{pbdb_subtaxa}
\alias{pbdb_subtaxa}
\title{pbdb_subtaxa}
\usage{
pbdb_subtaxa(data, do.plot, col)
}
\arguments{
\item{data}{data frame with our query to the 
paleobiodb \code{\link{pbdb_occurrences}}}

\item{do.plot}{by default this function make a plot to 
visualize the distribution of taxa. Set to FALSE to skip the plot.}

\item{col}{set the colour of the histogram. skyblue2 by default.}
}
\value{
a plot and a data frame with the number of subtaxa in the data.
}
\description{
Count the number of taxa (species, genera, families, orders,
classes, and phyla) in an occurrence data frame.
}
\examples{
\dontrun{
  canidae_quat <- pbdb_occurrences(
    limit = "all", base_name = "Canidae",  interval = "Quaternary",
    show = c("coords", "classext", "ident"), vocab = "pbdb"
  )
  pbdb_subtaxa(canidae_quat)
}

}
